Add unit tests for user notifications dispatcher to Detox. Tests cover the most common cases.

+++ detox/scripts/build.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
 BABEL_ENV=test babel src -d lib
+if [ "$1" != "noframework" ]; then
 echo -e "\nBuilding Detox.framework"
-xcodebuild build -project ios/Detox.xcodeproj -scheme DetoxFramework -configuration Release -derivedDataPath DetoxBuild > /dev/null
-cp -r DetoxBuild/Build/Products/Release-universal/Detox.framework .
-rm -fr DetoxBuild
-tar -cf Detox.framework.tar Detox.framework
+	xcodebuild build -project ios/Detox.xcodeproj -scheme DetoxFramework -configuration Release -derivedDataPath DetoxBuild > /dev/null
+	cp -r DetoxBuild/Build/Products/Release-universal/Detox.framework .
+	rm -fr DetoxBuild
+	tar -cf Detox.framework.tar Detox.framework
+fi
